Config CLI
* Moving all config items to entry_points and using click to create nested subcommands.
* Move helpers out of default import.
