+++ setup.cfg
@@ -44,8 +44,9 @@ setup_requires = pyscaffold>=3.2a0,<3.3a0
 install_requires =
     astroplan
     astropy
+    click
     Flask
-    google-cloud-bigquery[pandas,pyarrow]
+    google-cloud-bigquery[pandas, pyarrow]
     google-cloud-storage
     holoviews
     hvplot
@@ -58,7 +59,6 @@ install_requires =
     Pillow
     pyserial
     PyYAML
-    pyarrow>=0.17.1
     requests
     ruamel.yaml
     scalpl
@@ -91,7 +91,7 @@ testing =
 [options.entry_points]
 # Add here console scripts like:
 console_scripts =
-    panoptes-config-server = panoptes.utils.config.server:run
+    panoptes-config-server = panoptes.utils.config.cli:config_server_cli
 
 [test]
 # py.test options when running `python setup.py test`
@@ -187,6 +187,8 @@ exclude_lines =
 
 show_missing = True
 ignore_errors = True
+omit =
+    panoptes/utils/config/cli.py
 
 [pyscaffold]
 # PyScaffold's parameters when the project was created.
